@@ -1,14 +1,14 @@
 {
   "name": "core-js",
   "description": "Standard library",
-  "version": "1.1.1",
+  "version": "1.1.4",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/zloirock/core-js.git"
   },
   "main": "index.js",
   "devDependencies": {
-    "webpack": "1.11.x",
+    "webpack": "1.12.x",
     "LiveScript": "1.3.x",
     "grunt": "0.4.x",
     "grunt-cli": "0.1.x",
@@ -26,14 +26,14 @@
     "karma-phantomjs-launcher": "0.2.x",
     "karma-phantomjs2-launcher": "^0.3.2",
     "promises-aplus-tests": "2.1.x",
-    "eslint": "1.2.x"
+    "eslint": "1.3.x"
   },
   "scripts": {
     "grunt": "grunt",
     "lint": "eslint es5 es6 es7 js web core fn modules",
     "promises-tests": "promises-aplus-tests tests/promises-aplus/adapter",
-    "test": "npm run lint && npm run grunt client livescript karma:continuous && npm run promises-tests && lsc tests/commonjs",
-    "lt": "npm run lint && npm run grunt client livescript karma:lt && npm run promises-tests && lsc tests/commonjs"
+    "test": "npm run lint && npm run grunt livescript client karma:continuous library karma:continuous-library && npm run promises-tests && lsc tests/commonjs",
+    "lt": "npm run lint && npm run grunt livescript client karma:lt library karma:lt-library && npm run promises-tests && lsc tests/commonjs"
   },
   "license": "MIT",
   "keywords": [
@@ -56,13 +56,13 @@
     "Dict",
     "partial application"
   ],
-  "gitHead": "84dd361666804271a8fa2640478e1d1315c5343e",
+  "gitHead": "77902780b95f958650ec48b4dda934712952d630",
   "bugs": {
     "url": "https://github.com/zloirock/core-js/issues"
   },
   "homepage": "https://github.com/zloirock/core-js#readme",
-  "_id": "core-js@1.1.1",
-  "_shasum": "3ea0867d9bcd2076ef619e44075d3e7239453060",
+  "_id": "core-js@1.1.4",
+  "_shasum": "e0c75b4505ac3d2119db508b9a9ebabb368bfea3",
   "_from": "core-js@>=1.0.0 <2.0.0",
   "_npmVersion": "2.13.3",
   "_nodeVersion": "3.0.0",
@@ -71,10 +71,10 @@
     "email": "zloirock@zloirock.ru"
   },
   "dist": {
-    "shasum": "3ea0867d9bcd2076ef619e44075d3e7239453060",
-    "size": 338801,
+    "shasum": "e0c75b4505ac3d2119db508b9a9ebabb368bfea3",
+    "size": 340630,
     "noattachment": false,
-    "tarball": "http://registry.cnpmjs.org/core-js/download/core-js-1.1.1.tgz"
+    "tarball": "http://registry.npm.taobao.org/core-js/download/core-js-1.1.4.tgz"
   },
   "maintainers": [
     {
@@ -83,8 +83,8 @@
     }
   ],
   "directories": {},
-  "publish_time": 1440057671489,
-  "_cnpm_publish_time": 1440057671489,
-  "_resolved": "http://registry.cnpmjs.org/core-js/download/core-js-1.1.1.tgz",
+  "publish_time": 1441420610933,
+  "_cnpm_publish_time": 1441420610933,
+  "_resolved": "https://registry.npm.taobao.org/core-js/download/core-js-1.1.4.tgz",
   "readme": "ERROR: No README data found!"
 }
