@@ -7,18 +7,14 @@
     "esparse": "./bin/esparse.js",
     "esvalidate": "./bin/esvalidate.js"
   },
-  "version": "1.2.5",
+  "version": "2.6.0",
   "files": [
     "bin",
-    "test/run.js",
-    "test/runner.js",
-    "test/test.js",
-    "test/compat.js",
-    "test/reflect.js",
+    "unit-tests.js",
     "esprima.js"
   ],
   "engines": {
-    "node": ">=0.4.0"
+    "node": ">=0.10.0"
   },
   "author": {
     "name": "Ariya Hidayat",
@@ -32,27 +28,34 @@
   ],
   "repository": {
     "type": "git",
-    "url": "git+ssh://git@github.com/ariya/esprima.git"
+    "url": "git+https://github.com/jquery/esprima.git"
   },
   "bugs": {
-    "url": "http://issues.esprima.org"
+    "url": "https://github.com/jquery/esprima/issues"
   },
-  "licenses": [
-    {
-      "type": "BSD",
-      "url": "http://github.com/ariya/esprima/raw/master/LICENSE.BSD"
-    }
-  ],
+  "license": "BSD-2-Clause",
   "devDependencies": {
-    "jslint": "~0.1.9",
-    "eslint": "~0.4.3",
-    "jscs": "~1.2.4",
-    "istanbul": "~0.2.6",
-    "complexity-report": "~0.6.1",
-    "regenerate": "~0.5.4",
-    "unicode-6.3.0": "~0.1.0",
+    "codecov.io": "~0.1.6",
+    "escomplex-js": "1.2.0",
+    "eslint": "~1.3.1",
+    "everything.js": "~1.0.3",
+    "glob": "^5.0.14",
+    "istanbul": "~0.3.19",
+    "jscs": "~2.1.1",
     "json-diff": "~0.3.1",
-    "optimist": "~0.6.0"
+    "karma": "^0.13.3",
+    "karma-chrome-launcher": "^0.2.0",
+    "karma-detect-browsers": "^2.0.1",
+    "karma-firefox-launcher": "^0.1.6",
+    "karma-ie-launcher": "^0.2.0",
+    "karma-mocha": "^0.2.0",
+    "karma-safari-launcher": "^0.1.1",
+    "lodash": "^3.10.0",
+    "mocha": "^2.2.5",
+    "node-tick-processor": "~0.0.2",
+    "regenerate": "~0.6.2",
+    "temp": "~0.8.3",
+    "unicode-7.0.0": "~0.1.5"
   },
   "keywords": [
     "ast",
