@@ -1,6 +1,6 @@
 {
   "name": "webpack",
-  "version": "1.11.0",
+  "version": "1.12.1",
   "author": {
     "name": "Tobias Koppers @sokra"
   },
@@ -9,7 +9,7 @@
     "async": "^1.3.0",
     "clone": "^1.0.2",
     "enhanced-resolve": "~0.9.0",
-    "esprima": "^1.2.0",
+    "esprima": "^2.5.0",
     "interpret": "^0.6.4",
     "memory-fs": "~0.2.0",
     "mkdirp": "~0.5.0",
@@ -17,7 +17,7 @@
     "optimist": "~0.6.0",
     "supports-color": "^3.1.0",
     "tapable": "~0.1.8",
-    "uglify-js": "~2.4.13",
+    "uglify-js": "~2.4.24",
     "watchpack": "^0.2.1",
     "webpack-core": "~0.6.0"
   },
@@ -30,8 +30,8 @@
     "component-webpack-plugin": "~0.2.0",
     "coveralls": "^2.11.2",
     "css-loader": "~0.15.0",
-    "diff": "^1.4.0",
-    "eslint": "^0.24.0",
+    "diff": "^2.0.2",
+    "eslint": "^1.1.0",
     "eslint-plugin-nodeca": "^1.0.3",
     "express": "~4.13.1",
     "extract-text-webpack-plugin": "~0.8.0",
@@ -43,7 +43,7 @@
     "js-beautify": "^1.5.10",
     "json-loader": "~0.5.1",
     "less-loader": "^2.0.0",
-    "microtime": "^1.2.0",
+    "loader-utils": "^0.2.11",
     "mocha": "~2.2.0",
     "mocha-lcov-reporter": "0.0.2",
     "raw-loader": "~0.5.0",
@@ -61,9 +61,9 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+ssh://git@github.com/webpack/webpack.git"
+    "url": "git+https://github.com/webpack/webpack.git"
   },
-  "homepage": "http://github.com/webpack/webpack",
+  "homepage": "https://github.com/webpack/webpack",
   "main": "lib/webpack.js",
   "web": "lib/webpack.web.js",
   "bin": {
@@ -80,20 +80,20 @@
     "pretest": "npm run lint && npm run beautify-lint",
     "test": "mocha",
     "travis": "npm run cover -- --report lcovonly",
-    "lint": "eslint lib",
+    "lint": "eslint lib bin hot scripts",
     "beautify-lint": "node ./scripts/beautify-check",
     "beautify": "node ./scripts/beautify-rewrite",
     "precover": "npm run lint && npm run beautify-lint",
     "cover": "istanbul cover -x *.runtime.js node_modules/mocha/bin/_mocha",
     "publish-patch": "mocha && npm version patch && git push && git push --tags && npm publish"
   },
-  "gitHead": "5335781be8c06c2ff9443431a196f6a94ca80bb3",
+  "gitHead": "cfcac5e3b62a52012318204dc798d7394e9ac6a9",
   "bugs": {
     "url": "https://github.com/webpack/webpack/issues"
   },
-  "_id": "webpack@1.11.0",
-  "_shasum": "201002b637e254a4e356f5bd067725e6fc9a1924",
-  "_from": "webpack@*",
+  "_id": "webpack@1.12.1",
+  "_shasum": "532f8cb8d9ac9694f030581531c84dec349bd6c0",
+  "_from": "webpack@>=1.12.1 <2.0.0",
   "_npmVersion": "2.10.1",
   "_nodeVersion": "0.12.4",
   "_npmUser": {
@@ -107,14 +107,14 @@
     }
   ],
   "dist": {
-    "shasum": "201002b637e254a4e356f5bd067725e6fc9a1924",
-    "size": 105867,
+    "shasum": "532f8cb8d9ac9694f030581531c84dec349bd6c0",
+    "size": 106354,
     "noattachment": false,
-    "tarball": "http://registry.cnpmjs.org/webpack/download/webpack-1.11.0.tgz"
+    "tarball": "http://registry.npm.taobao.org/webpack/download/webpack-1.12.1.tgz"
   },
   "directories": {},
-  "publish_time": 1438860662053,
-  "_cnpm_publish_time": 1438860662053,
-  "_resolved": "http://registry.cnpmjs.org/webpack/download/webpack-1.11.0.tgz",
+  "publish_time": 1441283239229,
+  "_cnpm_publish_time": 1441283239229,
+  "_resolved": "https://registry.npm.taobao.org/webpack/download/webpack-1.12.1.tgz",
   "readme": "ERROR: No README data found!"
 }
