@@ -21,14 +21,14 @@
   },
   "homepage": "https://github.com/coocssweb/react-plug-in#readme",
   "dependencies": {
-    "babel-core": "^5.8.22",
+    "babel-core": "^5.8.24",
     "babel-loader": "^5.3.2",
     "jquery": "^2.1.4",
     "react": "^0.13.3",
-    "react-hot-loader": "^1.2.8",
+    "react-hot-loader": "^1.3.0",
     "react-router": "^0.13.3",
     "underscore": "^1.8.3",
-    "webpack": "^1.11.0"
+    "webpack": "^1.12.1"
   },
   "devDependencies": {
     "webpack-dev-server": "^1.10.1"
